--- src/main/java/org/gitlab4j/api/models/BuildGitStrategy.java
@@ -1,53 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2017 Greg Messner <greg@messners.com>
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to
- * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
- * the Software, and to permit persons to whom the Software is furnished to do so,
- * subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
- * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
- * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
- * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package org.gitlab4j.api.models;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonValue;
-import org.gitlab4j.api.utils.JacksonJsonEnumHelper;
-
-/**
- * Enum for the various Job status values.
- */
-public enum BuildGitStrategy {
-
-    FETCH, CLONE;
-
-    private static JacksonJsonEnumHelper<BuildGitStrategy> enumHelper = new JacksonJsonEnumHelper<>(BuildGitStrategy.class);
-
-    @JsonCreator
-    public static BuildGitStrategy forValue(String value) {
-        return enumHelper.forValue(value);
-    }
-
-    @JsonValue
-    public String toValue() {
-        return (enumHelper.toString(this));
-    }
-
-    @Override
-    public String toString() {
-        return (enumHelper.toString(this));
-    }
-}
